Project/backend/requirements.txt
absl-py==0.11.0
asgiref==3.2.3
astunparse==1.6.3
beautifulsoup4==4.9.1
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
cycler==0.10.0
defusedxml==0.6.0
dj-twilio-sms==2.1.0
Django==2.2.16
django-allauth==0.42.0
django-cors-headers==3.5.0
django-crontab==0.7.1
django-environ==0.4.5
django-rest-auth==0.9.5
django-sendsms==0.4
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
mysql-connector-python==8.0.20
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.13.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
absl-py==0.11.0
astunparse==1.6.3
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dlib==19.21.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0

